+++ ashley/src/test/kotlin/ktx/ashley/RigidBody.kt
@@ -1,0 +1,5 @@
+package ktx.ashley
+
+import com.badlogic.ashley.core.Component
+
+class RigidBody: Component